No error if node_modules doesn't exist

+++ build.ps1
@@ -1,5 +1,5 @@
 pushd dotnetcore-acquisition-library
-rm -r -force node_modules
+if (Test-Path node_modules) { rm -r -force node_modules }
 npm install
 npm run compile
 
@@ -11,7 +11,7 @@ if (! $?)
 popd
 
 pushd dotnetcore-acquisition-extension
-rm -r -force node_modules
+if (Test-Path node_modules) { rm -r -force node_modules }
 npm install
 npm run compile
 
@@ -23,7 +23,7 @@ if (! $?)
 popd
 
 pushd sample
-rm -r -force node_modules
+if (Test-Path node_modules) { rm -r -force node_modules }
 npm install
 npm run compile
